refactor(backend): extract god modules and fix dependency leak
- handlers/mod.rs: 2348→1375 lines (-41%)
  - Extract ExecEvent to events.rs
  - Extract AppError/ApiJson to errors.rs
  - Extract middleware (RequestId, propagate_request_id) to middleware.rs
  - Extract static file handlers, version/upgrade to static_handlers.rs

- db/mod.rs: 1745→1424 lines (-18%)
  - Extract usage stats CRUD to usage.rs

- services/message_debounce.rs: Fix dependency leak
  - Replace handlers::execution::start_todo_execution with direct
    executor_service::run_todo_execution calls
  - Service layer no longer depends on handler layer

957 tests pass.
